// apps/web/src/api/generated/model/providerConnectionDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */
import type { ProviderConnectionDtoType } from './providerConnectionDtoType.ts';

export interface ProviderConnectionDto {
  type: ProviderConnectionDtoType;
  id: string;
  name: string;
  baseUrl: string;
  hasApiKey: boolean;
  enabled: boolean;
  hiddenModelIds: string[];
  createdAt: string;
  updatedAt: string;
}