// apps/web/src/api/generated/model/updateProviderConnectionDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */

export interface UpdateProviderConnectionDto {
  /** @maxLength 80 */
  name?: string;
  /** @maxLength 2048 */
  baseUrl?: string;
  /** @nullable */
  apiKey?: string | null;
  enabled?: boolean;
  /**
     * @maxItems 1000
     * @items.maxLength 512
     */
  hiddenModelIds?: string[];
}